+++ addons/l10n_si_monthly_ajpes/wizard/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from . import generate_monthly_report_wizard